[jcommon] pom split
Summary:
per discussion w/Jason, the cleanest way to do things in jcommon is to have the root pom be an "aggregator" pom that is for convenience of running builds for all sub-directories.

The "parent" pom has been moved aside to the origin/pom.xml file.  It's a new version, and has been deployed

Test Plan:
1. mvn install  origin
2. mvn-test from root:

Reviewers: ehwang, groys, timw, wwang

Reviewed By: ehwang

CC: puma-dev@lists

Differential Revision: https://phabricator.fb.com/D451198