[jcommon] Implement QuantileDigest for quantile estimation
Summary:
This is a first cut at an implementation of Quantile Digests, as described by
http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.132.7343. A quantile digest is a data
structure for estimating quantiles in one pass over large streams of values using sublinear space,
with a guaranteed bound on the error of the estimation.

I'm looking for feedback before I make any further changes. This is a non-trivial data structure,
so I'd happy to give you the whiteboard overview before you dive into the code.

Other functionality that could be added later:
- compute the quantile for a given value
- count the number of values in a range (i.e., histogram-like)
- merging independently-computed digests (i.e., support for distributed estimation)
- getting multiple quantiles in batch (more efficient)
- exponential decay

This also looks promising as an extension that allows for more accuracy at the extremes:
http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.133.793

Finally, this class may be useful as a basis for implementing task #966148

Test Plan: Added unit tests.

Reviewers: srash, timw, dain

Reviewed By: srash

CC: jhoon, dain

Differential Revision: https://phabricator.fb.com/D456072

Task ID: 1024656