[jcommon] lock was held unnecessarily for decrement
Summary:
tim found in profiling that a lot of time was spent in a method that is called on every task execution in one of our layers over executors.  The lock was being held unnecessarily, and only for the purpose of notification when count was 0.

Now, only send notification when count is 0

Test Plan:
1. mvn -T 12 clean test
2. new rate test tests for reasonable rate, also outputs so we can track regression

Reviewers: timw, ehwang, wwang, groys, binu

Reviewed By: ehwang

CC: puma-dev@lists, datafreeway-dev@lists

Differential Revision: https://phabricator.fb.com/D480497