[jcommon] Array and LongArray
Summary:
1. introduce a simplified "list" or in this case, Array interface.  It's generic so that it can be used in classes that want to be generic, but allows for primitive-backed implementations

2. reference implementation of Array for Long.  There are data structures already implemented on long[] that if changed to use Array, can be changed to generic and then implementing Array for the type (one can use AbstractList possibly easier). Examples are the LongHashSet and a "heap" implementation for CTA that does 'flat' long-tuples.  Could be generified to any primtiive tuple

I made this thread-safe, mostly by synchronizing the methods.  Even the iterator *should* be thread-safe, but definitely does not throw CME.  it makes a best-effort to give as much valid data as it can.  It will not throw CME or give invalid data.  It's contract is similar to ConcurrentHashMap in that you are seeing a series of values from the list that were in it at sometime.

Ideal use of the iterator is NOT in a concurrent case, though, with some locking done to ensure this.

Thought: we could provide a SynchronizedSimpleArray that worked same as Collections.synchronizedX since SimpleArray doesn't specify thread-safety.

Test Plan:
1. existing tests
2. new unit tests for reference impl

Reviewers: timw, dain, martint, nitayj

Reviewed By: nitayj

CC: jcommon@, dphillips

Differential Revision: https://phabricator.fb.com/D506752