[jcommon] simple move of stats classes
Summary:
We have several topk implementations, and the puma approximate one will be moved in soon.  This creates a package for topk

It also fixes the poms so that if we want all poms to update their version (the effective version people use for any component, or jcommon-all),it's a single property in the parent pom.  It can still be overridden. in the parent and the library (ex: collections wants to be 2.0.0-SNAPSHOT, update its own pom version.  The parent may or may not choose to update the inherited version which means anyone using the lib might have to explicitly specify version 2.0.0-SNAPSHOT

This is good for experimental changes, or early RC testing of one component.

Test Plan: mvn clean test

Reviewers: timw, wwang, ajaiswal, groys

Reviewed By: groys

CC: puma-dev@lists

Differential Revision: https://phabricator.fb.com/D533522