[jcommon] Allow config to be pulled from a customized location (rather than only from files)
Summary:
Puma uses ConfigProviderImpl to parse config information, which uses a filed based parser ExpandedConfFileJSONProvider. Ideally we would like to support the config to be pulled from other locations (for example from a metastore). This is  to generalize ConfigProvderImpl to take a "ConfigResolver". The implementation is backward compatible.

Many thanks to Tim for providing feed back! I like the new way a lot more. :-)

Test Plan: mvn test

Reviewers: srash, timw, ajaiswal, ehwang

Reviewed By: timw

CC: security-diffs@lists

Differential Revision: https://phabricator.fb.com/D599175

Task ID: 1795673