[jcommon] Remove log4j dependency in code
Summary:
This diff tries to remove log4j dependency in code and let the library
only depend on slf4j-api. I also removed a lot of logging statements
in unit test because in my opinion unit tests should only print
when the they fail.

This diff only cleans up a minor bug in jcommon-concurrency, where
the file ExceptionHandler is put into a wrong package. This caused
the build in eclipse to fail.

Test Plan: mvn clean install in root directory.

Reviewers: srash, timw

CC: jaxlaw, security-diffs@lists

Differential Revision: https://phabricator.fb.com/D626976