[jcommon] Add capability to pass StatNameBuilder into JVMStatsExporter so that stats from Mbeans can be exported as fb303 with custom naming
JVMStatsExporter exports MBeans into fb303 stats without changing their names. Some of the MBean namings are non-intuitive (weird) like HBase Client uses weird looking "connection ID" into the name.
This change extends the JVMStatsExporter to receive a StatsNameBuilder object which gets a callback for deciding the name of the stat being exported. This builder can also act as a gatekeeper for stats export i.e if it returns null as name, the stat would not be exported.

Closes #44