Code cleanup for squid:S2293 impacting a variety of classes.
Java 7 introduced the diamond operator (<>) to reduce the verbosity of generics code. For instance, instead of having to declare a List's type in both its declaration and its constructor, you can now simplify the constructor declaration with <>, and the compiler will infer the type.

There was also one javadoc cleanup as well - the javadoc comment referenced a method argument that had changed its name. This was for method pack() in the class PackedByteArray.